Refactor job tasks to utilize utility functions for langfuseLangchainHandler and getInstallationIdFromRepositoryId.
