+++ frontend/packages/jobs/src/tasks/knowledge/generateSchemaMeta.ts
@@ -12,8 +12,8 @@ import { toJsonSchema } from '@valibot/to-json-schema'
 import { v4 as uuidv4 } from 'uuid'
 import { type InferOutput, boolean, object, parse, string } from 'valibot'
 import { SCHEMA_OVERRIDE_FILE_PATH } from '../../constants'
-import { langfuseLangchainHandler } from '../../functions/langfuseLangchainHandler'
 import { createClient } from '../../libs/supabase'
+import { langfuseLangchainHandler } from '../../utils/langfuseLangchainHandler'
 import { fetchSchemaInfoWithOverrides } from '../../utils/schemaUtils'
 import { createKnowledgeSuggestionTask } from './createKnowledgeSuggestion'
