+++ frontend/packages/jobs/src/tasks/review/generateReview.ts
@@ -5,8 +5,8 @@ import {
 } from '@liam-hq/github'
 import { logger, task } from '@trigger.dev/sdk/v3'
 import { v4 as uuidv4 } from 'uuid'
-import { langfuseLangchainHandler } from '../../functions/langfuseLangchainHandler'
 import { createClient } from '../../libs/supabase'
+import { langfuseLangchainHandler } from '../../utils/langfuseLangchainHandler'
 import { fetchSchemaInfoWithOverrides } from '../../utils/schemaUtils'
 import { generateReview } from './reviewPrompt'
 import { saveReviewTask } from './saveReview'
